+++ lib/src/extensions/form_field_state_extension.dart
@@ -90,28 +90,44 @@ extension FormFieldKeyExtension on Key {
     return null;
   }
 
-  /// Returns a [FieldKey] with this [Key] value.
-  FieldKey<T> field<T>({FieldAdapter<T>? adapter, bool? unmask}) {
-    final value = ArgumentError.checkNotNull(this.value, 'value');
-
-    return FieldKey(value, adapter: adapter, unmask: unmask);
-  }
-
-  /// Creates a `FieldKey<DateTime>` with this [Key] value.
-  FieldKey<DateTime> date([FieldAdapter<DateTime>? adapter]) {
-    return field(adapter: adapter);
+  /// Creates a [FieldKey] of type [T] of this [Key].
+  ///
+  /// - [adapter] is a function to adapt the [T] value to any value.
+  /// - [unmask] overrides [FormxOptions.unmask].
+  ///
+  /// The type [T] must be the same type defined in the associated [FormField].
+  /// This [Key] value must be a [String]. Otherwise, an [ArgumentError] is thrown.
+  FieldKey<T> field<T>({
+    FieldAdapter<T>? adapter,
+    bool? unmask,
+  }) {
+    return FieldKey(
+      ArgumentError.checkNotNull(value, 'value'),
+      adapter: adapter,
+      unmask: unmask,
+    );
   }
 
   /// Creates a `FieldKey<String>` with this [Key] value.
+  ///
+  /// Use with a [FormField] of type [String]. Ex: [TextFormField].
   FieldKey<String> text([FieldAdapter<String>? adapter]) {
     return field(adapter: adapter);
   }
 
-  /// Whether to unmask the value. Overrides [FormxOptions.unmask].
-  FieldKey<T> unmasked<T>() => field(unmask: true);
+  /// Creates a `FieldKey<DateTime>` of this [Key] value.
+  ///
+  /// Use with a [FormField] of type [DateTime]. Ex: [DateFormField].
+  FieldKey<DateTime> date([FieldAdapter<DateTime>? adapter]) {
+    return field(adapter: adapter);
+  }
 
-  /// Whether to mask the value. Overrides [FormxOptions.unmask].
-  FieldKey<T> masked<T>() => field(unmask: false);
+  /// Creates a `FieldKey<List<T>>` of this [Key] value.
+  ///
+  /// Use with a [FormField] of type [List]. Ex: [CheckboxListFormField].
+  FieldKey<List<T>> list<T>([FieldAdapter<List<T>>? adapter]) {
+    return field(adapter: adapter);
+  }
 }
 
 /// Extension for `FieldKey<T>`.
@@ -121,41 +137,95 @@ extension FieldKeyExtension<T> on FieldKey<T> {
 
   /// Whether to mask the value. Overrides [FormxOptions.unmask].
   FieldKey<T> masked() => copyWith(unmask: false);
+
+  /// Adapts [T] `value` to `Object.toMap()` in `FormState.values`.
+  ///
+  /// If [T] is an [Iterable], it will map each element to `Object.toMap()`.
+  FieldKey<T> toMap() {
+    return copyWith(adapter: (value) {
+      if (value is Iterable) return value.map((e) => e.toMap());
+      return (value as dynamic).toMap();
+    });
+  }
+
+  /// Adapts [T] `value` to `Object.toJson()` in `FormState.values`.
+  ///
+  /// If [T] is an [Iterable], it will map each element to `Object.toJson()`.
+  FieldKey<T> toJson() {
+    return copyWith(adapter: (value) {
+      if (value is Iterable) return value.map((e) => e.toJson());
+      return (value as dynamic).toJson();
+    });
+  }
 }
 
 /// Extension for `FieldKey<String>`.
 extension NumberFieldKeyExtension on FieldKey<String> {
-  /// Adapt [String] `value` to [int] in `FormState.values`.
+  /// Adapts [String] `value` to [int] in `FormState.values`.
   FieldKey<String> toInt() {
-    return copyWith(adapter: (value) => int.tryParse(value ?? ''));
+    return copyWith(adapter: int.tryParse);
   }
 
-  /// Adapt [String] `value` to [double] in `FormState.values`.
+  /// Adapts [String] `value` to [double] in `FormState.values`.
   FieldKey<String> toDouble() {
-    return copyWith(adapter: (value) => double.tryParse(value ?? ''));
+    return copyWith(adapter: double.tryParse);
+  }
+
+  /// Adapts [String] `value` to [Color] in `FormState.values`.
+  FieldKey<String> toColor() {
+    return copyWith(adapter: (value) {
+      if (value.startsWith('#')) value = value.substring(1);
+      if (value.length == 6) value = 'FF$value';
+      if (value.length != 8) return null;
+
+      final colorValue = int.tryParse(value, radix: 16);
+      return colorValue != null ? Color(colorValue) : null;
+    });
+  }
+}
+
+/// Extension for `FieldKey<String>`.
+extension EnumFieldKeyExtension on FieldKey<Enum> {
+  /// Adapts [Enum] `value` to [String] in `FormState.values`.
+  FieldKey<Enum> toName() {
+    return copyWith(adapter: (value) => value.name);
+  }
+
+  /// Adapts [Enum] `value` to [int] in `FormState.values`.
+  FieldKey<Enum> toIndex() {
+    return copyWith(adapter: (value) => value.index);
   }
 }
 
 /// Extension for `FieldKey<DateTime>`.
 extension DateFieldKeyExtension on FieldKey<DateTime> {
-  /// Adapt [DateTime] `value` to a [String] in `FormState.values`.
+  /// Adapts [DateTime] `value` to a [String] in `FormState.values`.
   FieldKey<DateTime> toLocalIso8601String() {
-    return copyWith(adapter: (value) => value?.toLocal().toIso8601String());
+    return copyWith(adapter: (value) => value.toLocal().toIso8601String());
   }
 
-  /// Adapt [DateTime] `value` to a [String] in `FormState.values`.
+  /// Adapts [DateTime] `value` to a [String] in `FormState.values`.
   FieldKey<DateTime> toUtcIso8601String() {
-    return copyWith(adapter: (value) => value?.toUtc().toIso8601String());
+    return copyWith(adapter: (value) => value.toUtc().toIso8601String());
   }
 
-  /// Adapt [DateTime] `value` to a [int] in `FormState.values`.
+  /// Adapts [DateTime] `value` to a [int] in `FormState.values`.
   FieldKey<DateTime> toMillisecondsSinceEpoch() {
-    return copyWith(adapter: (value) => value?.millisecondsSinceEpoch);
+    return copyWith(adapter: (value) => value.millisecondsSinceEpoch);
+  }
+}
+
+/// Extension for `FieldKey<List<T>>`.
+extension ListFieldKeyExtension<T> on FieldKey<List<T>> {
+  /// Adapts List [T] by mapping [toElement] in `FormState.values`.
+  FieldKey<List<T>> map(dynamic Function(T e) toElement) {
+    return copyWith(adapter: (value) => value.map(toElement).toList());
   }
 }
 
 /// A function to adapt a [Object] value to any value.
-typedef FieldAdapter<T extends Object?> = dynamic Function(T? value);
+typedef FieldAdapter<T extends Object?> = dynamic Function(T value);
+typedef FieldAdapterTo<T extends Object?, To> = To Function(T? value);
 
 /// A [Key] for a [FormField].
 @immutable
@@ -170,10 +240,10 @@ class FieldKey<T> extends GlobalObjectKey<FormFieldState<T>> {
   final FieldAdapter<T>? adapter;
 
   /// Adapts the [value] if same type. Otherwise, returns the same [value].
-  dynamic maybeAdapt(dynamic value) {
+  dynamic maybeAdapt(Object value) {
     if (value is! T) return value;
 
-    return adapter?.call(value);
+    return adapter?.call(value as T) ?? value;
   }
 
   /// Whether to unmask the value. Overrides [FormxOptions.unmask].